RFC: Use different exit statuses for different failures
(This isn't obviously the right approach, so feedback welcome.)

The problem that I'm having is that I get occasional image pull "failures." That is to say - these are not persistent, retrying them usually fixes things.

I'd like to automate the retries but the current retry code makes that difficult (1 is too generic). `docker run` returns 125 for a missing image. This patch makes buildkite return the same.